+++ ksp/sealed/generator/test/src/commonMain/kotlin/Test.kt
@@ -7,10 +7,10 @@ import dev.inmo.micro_utils.ksp.sealed.GenerateSealedWorkaround
 @GenerateSealedTypesWorkaround
 sealed interface Test {
     @GenerateSealedWorkaround.Order(2)
-    @GenerateSealedTypesWorkaround.Order(2)
+    @GenerateSealedTypesWorkaround.Exclude
     object A : Test
     @GenerateSealedWorkaround.Exclude
-    @GenerateSealedTypesWorkaround.Exclude
+    @GenerateSealedTypesWorkaround.Order(2)
     object B : Test
     @GenerateSealedWorkaround.Order(0)
     @GenerateSealedTypesWorkaround.Order(0)