+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.24.3
 
+* `Ksp`:
+  * `Sealed`:
+    * Fixes in processing of `GenerateSealedTypesWorkaround` annotations
+
 ## 0.24.2
 
 * `Versions`: